refactor: simplified projects resource classes packaging
