+++ knowledge-graph/src/test/scala/io/renku/knowledgegraph/projects/GitLabProjectFinderSpec.scala
@@ -16,8 +16,9 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.projects.rest
+package io.renku.knowledgegraph.projects
 
+import ProjectsGenerators._
 import cats.effect.IO
 import cats.syntax.all._
 import eu.timepit.refined.api.Refined
@@ -27,16 +28,15 @@ import io.circe.Json
 import io.circe.literal._
 import io.renku.generators.CommonGraphGenerators.accessTokens
 import io.renku.generators.Generators.Implicits._
-import io.renku.graph.model
 import io.renku.graph.model.GraphModelGenerators.projectPaths
+import io.renku.graph.model.projects
 import io.renku.http.client.RestClient.ResponseMappingF
 import io.renku.http.client.{AccessToken, GitLabClient}
 import io.renku.http.server.EndpointTester._
 import io.renku.interpreters.TestLogger
+import io.renku.knowledgegraph.projects.GitLabProjectFinder.GitLabProject
 import io.renku.knowledgegraph.projects.model.Permissions
 import io.renku.knowledgegraph.projects.model.Permissions._
-import io.renku.knowledgegraph.projects.rest.GitLabProjectFinder.GitLabProject
-import io.renku.knowledgegraph.projects.rest.ProjectsGenerators._
 import io.renku.stubbing.ExternalServiceStubbing
 import io.renku.testtools.{GitLabClientTools, IOSpec}
 import io.renku.tinytypes.json.TinyTypeEncoders._
@@ -59,7 +59,7 @@ class GitLabProjectFinderSpec
   "findProject" should {
 
     "return fetched project info if service responds with OK and a valid body" in new TestCase {
-      forAll { (path: model.projects.Path, accessToken: AccessToken, project: GitLabProject) =>
+      forAll { (path: projects.Path, accessToken: AccessToken, project: GitLabProject) =>
         val expectation = project.some
 
         (gitLabClient
@@ -105,7 +105,7 @@ class GitLabProjectFinderSpec
     implicit val gitLabClient: GitLabClient[IO] = mock[GitLabClient[IO]]
     val projectFinder = new GitLabProjectFinderImpl[IO]
 
-    def uri(path: model.projects.Path) = uri"projects" / path.show withQueryParam ("statistics", "true")
+    def uri(path: projects.Path) = uri"projects" / path.show withQueryParam ("statistics", "true")
 
     val endpointName:         String Refined NonEmpty = "single-project"
     implicit val accessToken: AccessToken             = accessTokens.generateOne
